requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-Limiter>=3.0
Flask-CORS>=4.0
Authlib>=1.0
python-dotenv>=1.0
requests>=2.31
PyMySQL>=1.0
reportlab>=4.0
Pillow>=10.0
PyJWT>=2.8
qrcode>=7.4
weasyprint>=60.0
cffi>=1.15.0